# api/v3/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: miniyu97@gmail.com
    name: meteormin
    url: https://github.com/meteormin/friday.go
  description: Friday.go API
  title: Friday.go API
  version: "{{.Version}}"
servers:
- url: "http://{{.Host}}/"
- url: "https://{{.Host}}/"
...
paths:
...
  /api/posts:
    get:
      description: 포스트 리스트 조회 API
      operationId: posts.retrieve
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/internal_adapter_rest_handler.PostResource'
                type: array
          description: 포스트 리스트 조회 성공
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      security:
      - BearerAuth: []
      summary: 포스트 리스트 조회
      tags:
      - posts
    post:
      description: 포스트 생성 API
      operationId: posts.create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapter_rest_handler.CreatePostRequest'
        description: 포스트 생성 정보
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapter_rest_handler.PostResource'
          description: 포스트 생성 성공
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 잘못된 요청
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 이메일 중복
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      security:
      - BearerAuth: []
      summary: 포스트 생성
      tags:
      - posts
      x-codegen-request-body-name: req
...
  /api/sites:
    get:
      description: 사이트 리스트 조회 API
      operationId: sites.retrieve
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/internal_adapter_rest_handler.SiteResource'
                type: array
          description: 사이트 리스트 조회 성공
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 잘못된 요청
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      security:
      - BearerAuth: []
      summary: 사이트 리스트 조회
      tags:
      - sites
    post:
      description: 사이트 생성 API
      operationId: sites.create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapter_rest_handler.CreateSiteRequest'
        description: 사이트 생성 정보
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapter_rest_handler.SiteResource'
          description: 사이트 생성 성공
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 잘못된 요청
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 이메일 중복
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      security:
      - BearerAuth: []
      summary: 사이트 생성
      tags:
      - sites
      x-codegen-request-body-name: req
  /api/sites/{id}:
    delete:
      description: 사이트 삭제 API
      operationId: sites.delete
      parameters:
      - description: 사이트 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapter_rest_handler.SiteResource'
          description: 사이트 삭제 성공
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 잘못된 요청
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 사이트 없음
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      security:
      - BearerAuth: []
      summary: 사이트 삭제
      tags:
      - sites
    get:
      description: 사이트 조회 API
      operationId: sites.find
      parameters:
      - description: 사이트 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapter_rest_handler.SiteResource'
          description: 사이트 조회 성공
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 잘못된 요청
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 사이트 없음
      security:
      - BearerAuth: []
      summary: 사이트 조회
      tags:
      - sites
    put:
      description: 사이트 수정 API
      operationId: sites.update
      parameters:
      - description: 사이트 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapter_rest_handler.UpdateSiteRequest'
        description: 사이트 수정 정보
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapter_rest_handler.SiteResource'
          description: 사이트 수정 성공
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 잘못된 요청
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 사이트 없음
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      security:
      - BearerAuth: []
      summary: 사이트 수정
      tags:
      - sites
      x-codegen-request-body-name: req
  /api/sites/{id}/posts:
    get:
      description: 사이트 포스트 리스트 조회 API
      operationId: sites.retrievePosts
      parameters:
      - description: 사이트 ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/internal_adapter_rest_handler.PostResource'
                type: array
          description: 사이트 포스트 리스트 조회 성공
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 잘못된 요청
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 사이트 없음
      security:
      - BearerAuth: []
      summary: 사이트 포스트 리스트 조회
      tags:
      - sites
  /api/upload-file:
    post:
      description: 파일 업로드 API
      operationId: files.upload
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/files_upload_request'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapter_rest_handler.UploadFileResponse'
          description: 파일 업로드 성공
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 잘못된 요청
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      security:
      - BearerAuth: []
      summary: 파일 업로드
      tags:
      - upload-file
  /api/upload-file/{id}:
    get:
      description: 파일 다운로드 API
      operationId: files.download
      parameters:
      - description: 파일 ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                items:
                  type: integer
                type: array
          description: 파일 다운로드 성공
        "400":
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 잘못된 요청
        "404":
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 파일 없음
        "500":
          content:
            application/octet-stream:
              schema:
                $ref: '#/components/schemas/github_com_meteormin_friday_go_internal_app.Error'
          description: 서버 오류
      security:
      - BearerAuth: []
      summary: 파일 다운로드
      tags:
      - upload-file
...
components:
  schemas:
    github_com_meteormin_friday_go_internal_app.Error:
      example:
        code: 0
        message: message
        title: title
      properties:
        code:
          type: integer
        message:
          type: string
        title:
          type: string
      type: object
    internal_adapter_rest_handler.CreatePostRequest:
      description: 생성 요청
      properties:
        content:
          type: string
        fileId:
          type: integer
        siteId:
          type: integer
        tags:
          items:
            type: string
          type: array
        title:
          type: string
      type: object
    internal_adapter_rest_handler.CreateSiteRequest:
      properties:
        host:
          type: string
        name:
          type: string
      type: object
    internal_adapter_rest_handler.PostResource:
      example:
        updated_at: updated_at
        created_at: created_at
        id: 6
        title: title
        uri: uri
        content: content
        fileId: 0
        tags:
        - tags
        - tags
      properties:
        content:
          type: string
        created_at:
          type: string
        fileId:
          type: integer
        id:
          type: integer
        tags:
          items:
            type: string
          type: array
        title:
          type: string
        updated_at:
          type: string
        uri:
          type: string
      type: object
...
    internal_adapter_rest_handler.SiteResource:
      example:
        createdAt: createdAt
        host: host
        name: name
        id: 0
        posts:
        - updated_at: updated_at
          created_at: created_at
          id: 6
          title: title
          uri: uri
          content: content
          fileId: 0
          tags:
          - tags
          - tags
        - updated_at: updated_at
          created_at: created_at
          id: 6
          title: title
          uri: uri
          content: content
          fileId: 0
          tags:
          - tags
          - tags
        updatedAt: updatedAt
      properties:
        createdAt:
          type: string
        host:
          type: string
        id:
          type: integer
        name:
          type: string
        posts:
          items:
            $ref: '#/components/schemas/internal_adapter_rest_handler.PostResource'
          type: array
        updatedAt:
          type: string
      type: object
...
    internal_adapter_rest_handler.UpdateSiteRequest:
      properties:
        name:
          type: string
      type: object
    internal_adapter_rest_handler.UploadFileResponse:
      example:
        id: 0
        uri: uri
      properties:
        id:
          type: integer
        uri:
          type: string
      type: object
...
    files_upload_request:
      properties:
        file:
          description: 파일
          format: binary
          type: string
      required:
      - file
      type: object
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...